[Core] Add User to hasAccess signature
This updates the signature for hasAccess to include the
User whose access is being checked as an explicit parameter.

Nearly every access check previously started with User::singleton
to get the logged in user (because, obviously, it's the user whose
access is being checked) before this change.

This updates the signature to take the user as a parameter and
remove the singleton calls.