[API] Displays QC info of MINC images, and allow hyphens in MINC filenames (#3670)
The API was not displaying QC info for MINC images because it was trying to find the file Qc.php; however the file is called QC.php. So change this in the .htaccess file to be consistent with the codebase filenaming or allowing MINC files to have hyphen in them (which might happen if the Loris-MRI project had a hyphen in it).

This fixes those bugs.