Code Issues 15 Pull requests 73 Projects 4 Wiki Insights Settings [Core] .htaccess - Redirect base URL requests to be routed through index.php (#3768)
Most existing LORIS instances have their apache configurations set so that
main.php is the entry point for all HTTP requests as was instructed in a previous release.

As of 20.0, .htaccess was modified so that all everything gets routed through index.php so long as the REQUEST_FILENAME is not a findable file by apache. This results in a REQUEST_FILENAME that is equal to the full path of main.php whenever there is no specifier placed after the LORIS base URL which causes the RewriteCond to fail and default the request through main.php

This is problematic for login & dashboard since they are home pages, and ultimately results in their setup functions never getting called / tpl_data never getting assigned:

This modification to .htaccess allows users to upgrade without having to modify their apache configurations, however users should also remove the `DirectoryIndex` directive from their apache config so that this is not required.