cleanup console output when CouchDB not configured (#3630)
This will prevent console output and when the CouchDB port not being a number or the host hasn't been configured.

The user may not have configured the port after installing Loris. Fixed by preventing the error message when the function _getRelativeURL occurs inside the CouchDB.class.inc class. A LorisException is thrown to show the error and no console warning is displayed.

See also: Bug #14383