[Instrument Manager] (20.0) Fix bug where LorisException should have been thrown when quatUser was not properly configured, improve error checking (#3746)
This addresses two related issues in Redmine.

- Correctly check if automatic instrument installation is possible by using empty rather than isset. Previously this condition would be true even with blank credentials in the config file.
- Dynamic feedback has been integrated to inform users of various differing server (mis)configurations.

Previously user's were informed when uploading had been "disabled by an administrator" (ie. vital project/ folders were not writable). In addition, users are now told that automatic installation is "disabled" when quatUser is not properly configured.

When an upload fails for any number of reasons, users are now informed as to why and how to fix the problem. In cases where a back-end administrator is needed to remedy an issue, additional information will be entered in the error log.