[Candidate] Sanitize candidate comments (#3834)
Before injecting values from the DB into the front-end HTML, they must be sanitized to prevent cross-site scripting (XSS) attacks. To sanitize on output of the information - instead of on input to the DB - is considered best-practice in the context of XSS attack mitigations (as opposed to SQLi prevention, for example).