[survey module] Remove checks for permissions when entring direct data (#3848)
This implements a new function to be used solely to load survey instruments. this function is based on the regular loadinstrument() function in the caller but removes all unecessary calls pertaining to regular instrument loading whn logged in. Surveys is an anonymous module and thus should not be checking for permissions or loading panels...