[mri_violations] Make 'Resolution Status' select element have auto width
created a new css file for this module because I felt that it didn't fit anywhere else in main.css or bootstrap css. let me know if you feel it should go somewhere else.

there is also a way to make it so that the select element box wraps around each select option (select element width becomes the width of the text for each selected option) but it would require adding to the js files of this module which are already really messy, so i would rather not. but let me know if this added styling is necessary.