[Help Editor] Use safe update instead of unsafe update (#3866)
The HelpFile class has been changed to use the safe update function from the database class instead of the explicitly unsafe version.

unsafeupdate is to be used with caution, only when you know you need to insert HTML and know that you can trust it.  This change fixes a security issue while preserving the markup functionality of the help editor. HTML will not directly render but Markdown will.