[Core] Remove PT Sans and PT Sans Bold fonts (#3901)
This removes the PT Sans and PT Sans Bold fonts
from the htdocs directory (and the reference to
them in CSS.) They're only used in one place,
released under a different license from LORIS,
and only used for public modules. The overhead of
keeping them in LORIS is probably not worth the
benefit of using them over the generic sans-serif
font family.