[Core: Database] Add exception on fetchAll failure case (#3880)
In an effort to take advantage of PHP's type system, we should add type declarations to our code, especially in the core classes.

The PDO fetchAll returns the empty set when no data in the DB matches a given query and FALSE on failure. However we do not distinguish this in our codebase and so a FALSE value is allowed to bubble up through subsequent DB calls.

This makes it so that the database class throws an exception in the failure case following discussion in #3854. This will allow us to remove Bool as one of the return types in the future.

Related issues: #3854, #3878