[Tools/Normalize_consent] Remove validation for consent status "no" (#3914)
Initially, we assumed that consent date would only be given if status=yes. Therefore, if status=no but there is a consent date, it is assumed that consent was withdrawn, thus, a withdrawn date is required. The script validates there being a withdrawn date for withdrawn consent. However, some projects also collect a consent date for status=no, breaking down our assumption on withdrawal. This PR removes that validation.