[Instrument Manager] Proof of concept load instrument manager from external repository
This updates the instrument manager to be loaded from a separate git
repo and composer required. Theoretically, this allows us to more
rapidly iterate on it, and make releases of the instrument manager
separately from LORIS releases to get module bug fixes / features out
faster by running "composer update".

The instrument manager is used as a test case for this because it is
relatively self contained, lightly used, and doesn't have any ajax
that needs to be converted to a router.

LORIS (including the instrument manager) should work identically
before and after this change. To test this PR, test the instrument
manager. Note that you may need to change the "vcs" line to a local
repository and tag a v21.0.0 to test this PR, because there's a
bootstrapping problem where the module requires >20, which isn't
released yet.

I've used created/used the loris2 github organization as a test
for this (loris was already taken), this is not necessarily its
final home (but making a aces repo for every single module is
probably not feasible.)