[Core: Login, User Accounts] Allow certain users to access the system in a time window basis (#3898)
Add ability to give users access to the system for a predefined time frame. "Active From" and "Active To" columns are added in user_accounts, which define a window in which the account may be used. This allows studies to either pre-create accounts that aren't active immediately, or expire accounts (or both.)

Thanks to CHU Sainte-Justine for this feature.