Add instruments module to load instruments through new router
This adds a new "instruments" module which takes the place of
what was previously done by NDB_Caller and main.php to load
instruments. The module does not do anything but delegate to
existing instruments that are in the project directory.

To test this: test a variety of instruments after this change,
in both LINST and PHP format. There should be no noticeable
difference.