Add type check to HTMLEscapeArray
In the process of adding return types and type declarations to the
Database class, the LORIS integration tests failed when an int value was
passed to htmlspecialchars.
This PR checks that inputs to that function are of string type before
paassing them as arguments.
A comment is added to this function to let developers know that
htmlspecialchars as called in the DB is not a magic bullet to preventing
JS injection.