Add rate-limting mechanism for log ins.
Set a sliding window for which users may be locked out if they make too
many bad login attempts. Two windows are used to balance a concern for
security and useability. E.g. for now a user is allowed to make 10 bad
attempts within 15 minutes and 15 attempts within an hour.
Do happy path refactoring for the passwordAuthenticate logic.