[MRI Preprocessing]: Adding tables to support BIDS dataset creation on the MRI side (#3912)
Add tables to the database to support the creation of a BIDS dataset for imaging data.
This is mostly to be used with PR #198 (aces/Loris-MRI#198) on the MRI side.

Essentially 2 new tables are created, bids_mri_scan_type_rel and bids_category.
The former links to the mri_scan_type table and creates a link between the existing acquisition protocol IDs, and all the related derivatives needed from it to create a BIDS compliant dataset/filenames. The latter defines the different BIDS categories which structures the different MRI modalities in sub-directories.

More specifically, a detailed column description per table is as follows:

MRIScanTypeID: link to the mri_scan_type ID column (e.g. 44 for t1 images in our default schema)
BIDSCategory: The values inside this tables are used to create the appropriate folder in which to house the nii images. For example, a t1 scan would be in a folder called anat because it fits in the anatomical category.
BIDSScanTypeSubCategory; this is mostly for fMRI. These func acquisitions need extra measures to define them (that also get stored in the accompanying JSON header file), so for a resting state fMRI, files should have a task-rest identifier, or what I define as the BIDSScanTypeSubCategory. A memory task fMRI would have a value of task-memory in the BIDSScanTypeSubCategory column. Other acquisitions can leave this field blank
BIDSScanType is the mri_scan_type Scan_type column, but named in a BIDS compliant manner. So t1 becomes T1w, dti becomes dwi, etc... this is the column that would rename our schema's Scan_type to whatever BIDS wants it to be
BIDSMultiEcho is a column that specifies which modalities are acquired in a multi-echo fashion. This is needed because those files need to be identified in BIDS with the string echo-1, echo-2, etc... in their filenames, so fill in this information in this column.

The MRI script in the referenced PR above needs to be run after the patches here have been applied.