[Database] Make pselectRow more exceptional (#3910)
This makes pselectRow error out in more invalid cases. In particular,
it throws an exception if called on a row that returns more than
one row (and as a result, also adds this behaviour to pselectOne) since
there's no way to know which row was intended. It also returns null,
not an empty array, if no rows are matched.

This improves LORIS by making latent errors more obvious
and exposing problems that might otherwise go unnoticed.