[JSX] Add new FilterableDataTable component
This adds a new FilterableDataTable component in order to make it easier to
create tables which have filters that dynamically restrict the content.

The filterable datatable component is like a DynamicDataTable, except it also
creates a filter based on a object passed to it as a prop. The data in the table
and the filters are dynamically linked.