[API] Add new API module and begin transition
This adds a new API module to replace the ad-hoc PHP
scripts and modrewrite rules which currently make up the
LORIS API. The module currently handled the login endpoint
only, but does so for all versions of the API rather than
having a separate handler for each.

The mod_rewrite rules are updated such that the login
endpoint is delegated to the module, while other endpoints
fall back on the older implementations to help the transition.

To test this try using the API login endpoint, as well as
an endpoint which is not handled by the module. Both should
work.