[candidate_list] Convert ajax script to real LORIS endpoint
This converts the validateProfileIDs AJAX script from an ajax
dir hack into a real LORIS endpoint. It also makes the page
decoration middleware slightly smarter so that it only decorates
pages where the Content-Type header is either empty or "text/html"
so that things that are explicitly not HTML don't have the HTML
page decorations added.