[create timepoint] Fix error in logic when projects are enabled (#3943)
If projects are disabled all subrojects should be displayed by default when creating a timepoint (this was and still is working). When projects are enabled and the project_rel has data for all projects only the designated subprojects should be displayed in the dropdown when creating a timepoint (this was and still is working). When projects are enabled but the project_rel table IS NOT populated, no options should be displayed when creating a timepoint; this fixes that by adding a message asking the administrator to configure projects for the LORIS instance.