[data_release] Fix access for admin (#3975)
The data_release test plan mentions that "2. Make sure that only the 'superuser' can ... as well as the list of already uploaded files" which makes sens but is not currently the case. This changes the query slightly to make sure the Admin user can see all the files and only regular users have a limited subsection.