Add document describing new routing (#3997)
This adds a document to the docs directory explicitly telling developers not to depend on the module/*/ajax directory and puts in place a (commented out, until more modules are updated) deprecation warnings in AjaxHelper.php.

The document explains why the deprecation is happening and points to an example PR of how to update an ajax script to the new architecture. It can be updated to be more helpful as people have more experience doing so or find common patterns, but for now this is better than nothing.