[Instrument] Fix warning by changing to "continue 2" (#4008)
A line in the NDB_BVL_Instrument class was generating the warning:

PHP Warning: "continue" targeting switch is equivalent to "break". Did you mean to use "continue 2"? in /var/www/loris/php/libraries/NDB_BVL_Instrument.class.inc on line 2118
on version 7.3

"continue 2" or "break" would result in the same behavior in this instance since the loop only contains a switch, this changes it to "continue 2" because that seems to be closer with what the original author had in mind.

For more info see: http://php.net/manual/en/control-structures.continue.php