[Core: User] Fixed Type Check for hasCenter Method (#4020)
The hasCenter method enforces strict type checking, however in the function doc it explicitly says that it expects an int. This int is then forced to be compared with a string provided by the getCenterIDs method. Therefore, if an int is passed to the function, the result will necessarily be false without this change.