[JSX] Dicom Archive - Refactor and Clean-Up (#4097)
This refactors the Dicom Archive JSX to meet the latest LORIS React standards.

It absorbs the changes made in PR #3956. It is also in accordance with the standards of the template merged from PR #4054. In addition, it adds the ability to configure a 'date' element in the context of the Menu Filter.