[Core: Utility] Deprecate Utility::toArray() (#4151)
Following #4133 toArray() is being deprecated in favour of the more precise function associativeToNumericArray(). This function does the same thing as toArray() but only accepts and returns arrays.

This change reduces confusion between toArray and asArray as well as allows us to be more precise with the way we're using types.

The calling code has been updated to use the new function and issue a deprecation warning.