[Core] Add Guzzle dependency (#4188)
Guzzle is an implementation of a PSR18 Http Client. It is used to make requests to external servers. As discussed in a recent LORIS meeting, this is required for many different functions we would like to include in LORIS.

Guzzle is chosen as it is the most-downloaded PSR18 implementation and likely to be well-maintained in the foreseeable future, but can be swapped out in the future.

Adding this dependency will allow for different features that are currently blocked to be developed.

Fixes #4187.