[JSX] Added Modal Window Component (#4253)
Added a Modal Window and Triggerable Modal Window Component to LORIS.

This allows developers to integrate forms right into the page, essentially providing a visual portal for the users to interact in without having to navigate anywhere. This presents itself as a more elegant solution than using a bootstrap modal window because it's fully React and can be customized in the future if we so desire!

Modal Window can be used as standalone component which is passed an open prop that dictates whether or not it is displayed.

A Triggerable Modal Window can be used to abstract away the responsibility of controlling the display of the modal window. A TriggerableModal will render a trigger button that, when clicked, will cause the Modal Window to be displayed.