[phan] Remove redeclaration of PHP internal getallheaders function
This removes the intentional redeclaration of the PHP getallheaders
function in SinglePointLogin.

It was previously required because getallheaders() only existed under
Apache. According to the PHP documentation (http://php.net/manual/en/function.getallheaders.php)
it became available under the CLI server in PHP 5.5.7, so it should
no longer be required for Travis or development purposes under
supported versions of PHP.