[Core] Fix TypeError in Notifier (#4323)
A previous change introduced a TypeError issue by using the PHPDoc return type as two functions' return type when the functions were not annotated properly.

This removes one function altogether as it acted as a wrapper for \User which can be directly accessed. The other was changed to use the array type instead of the string type.