[phan] Fix error about continue targetting switch
Fix errors in instrument manager about a continue statement inside
of a switch statement. These are changed to a break, since that's
how the continue behaves in that context anyways. "continue 2" or
"break" are equivalent in this context, since the switch in question
is at the end of a loop, so breaking out of the switch will already
continue to the next iteration.