[Brainbrowser] Fix images not loading after fixing paths in violation tables (#4347)
The paths stored in the 3 violation tables (MRICandidateErrors, mri_violations_log and mri_protocol_violated_scans) were wrong and a tool script has been provided on the LORIS-MRI side to correct the paths to the MINC files so that they are valid paths.

Due to the bad paths being present in those tables, brainbrowser was hacked on the LORIS side to be able to display the images. Now that the paths are corrected, this hack can be removed which also simplifies the code.