[Login] Fix issue where password reset emails were not sent (#4362)
This changes the module to use the factory function instead of the singleton function from the User class. Singleton is not appropriate as the user requesting the password reset has not yet logged in, resulting in it returning an AnonymousUser instance. This prevents the retrieval of the Email value for users attempting to reset their passwords.