[Login: Request Account] Fixing 500 Error that arises due to email discrepancy  (#4372)
This avoids a 500 server error that happens on the request account page when there's a discrepency between the UserID and Email address provided and an existing account. It adds a hack where the user can not request an account when there's a UserID *or* Email address of an existing account that has the same email address as that provided by the user.

While this further conflates the two concepts that shouldn't be used interchangeably, for the moment it avoids a 500 error.