[Instruments] Add ability to load instrument JSON data from flag Data column. (#3894)
LORIS previously saved instrument data to a "Data" column
in the flag table as a test, this change gives users the
option to experiment with using it. Loading data from the
Data column means that the instrument SQL tables are no
longer required, but should be considered experimental and
not ready for production.

Use this, set the "$jsonData" variable in an instrument
to true, and data should be saved and loaded from the
flag table instead of a separate instrument table.

This feature is still experimental and not recommended for
production use with sensitive data.