The NDB_BVL_Instrument class has two functions which select from the database to get metadata about the instrument: getFullName and getSubtestList.
Neither of these are pieces of metadata that should
be coming from the database, since they're information
about the instrument itself, not data.

This makes the functions abstract, forcing instruments
to implement the functions themselves and return appropriate
values.

A trait is provided to provide the legacy behaviour for
existing instruments.