[Database] Add NOQUOTES as an HTML escaping option for JSON (#4374)
Modify the unsafe* family of functions to allow for a NOQUOTES option
when inserting data. This provides a middle ground between escaping all
HTML characters and none. This is important for working with JSON in the
database as escaping the quotation marks seems to make the data
difficult to work with.