[Core] Add pwned passwords API to prevent common passwords and password reuse (#3961)
Add the pwned passwords API into the LORIS password logic. This is part of a refactoring effort to modernize the way LORIS does passwords.

In particular, this API prevents users from creating passwords that are known to be present in previous data leaks. After a password breach, hackers compose large lists of the password and try them in brute force attacks against the sites.

With this feature users will be prevented from creating a password like Spring2016! which is a bad password (because it's in a public corpus of passwords known to have been compromised in previous breaches) even though it passes the (current) LORIS password strength requirements.