[imaging] Add SQL patch to add image_type to mri_protocol table (#4448)
This contains the SQL patches to run to add the image_type check in the mri_protocol and mri_protocol_violated_scans.

This image_type header field is a key variable as this is the only MINC header that allows to differentiate between a fieldmap phase difference from a fieldmap magnitude file which have different purposes when it comes to analyses and the insertion pipeline should be able to recognize one from the other. This is also the case for MP2RAGE modalities and other modern imaging modalities, hence the need to add that field in the mri_protocol table.

Example:

    image_type for fieldmap phase difference image: ORIGINAL\\PRIMARY\\P\\ND (P for phase)
    image_type for fieldmap magnitude image: ORIGINAL\\PRIMARY\\M\\ND (M for magnitude)

Note that because of the multiple \ in that field, in order to insert the correct value, each \ needs to be escaped. Example:
INSERT INTO mri_protocol SET image_type='ORIGINAL\\\\PRIMARY\\\\P\\\\ND', ...

See also LORIS MRI PR: aces/Loris-MRI#416