[jsx/DataTable] Fix reversions in data table UI (#4629)
The issue tracker PR which reverted changes to the data table also caused some regressions to the UI that should not have been undone. This restores the new UI and removes the duplicated "Download as CSV" button which is now placed in the call to action section of the table.