[statistics] fix imaging tab loading when Scan_Done is not in mri_parameter_form (#4672)
The imaging tab of statistics depends on the scan types defined in the mri_scan_type table. each scan type is crossmatched with a column in the mri_parameter_form table. if that column does not exist the SQL query is now bypassed and an error message is displayed stating that the column does not exist