[login] check for X-FORWARDED-FOR before REMOTE_ADDR (#4699)
The user_login_history table of the database has been logging the IP address of the proxy instead of the real IP of the user. To prevent that, we are checking for an IP address set in the X-FORWARDED-FOR variable and only if that is NULL we are fetching the REMOTE_ADDR and setting it.