[instrument_list] Fix BVLQCType on ControlPanel (#4683)
Fix bug where when instrument data entry and DDE is complete with no conflicts "Send to DCC" is not available, even with Send to DCC permission and visit set to Pass, and clicking on any box for BVL QC (e.g. Visual, Hardcopy) does not register nor does clicking Complete for BVL QC Status as a result.