[Issue Tracker] Fix incorrect links to individual issues (#4763)
There was a regression introduced in #4112 where the issue tracker
links reverted to getting the ID from a query parameter instead
of the URL. This was done inconsistently, and as a result some
places loaded URLs such as "...&issueID=?issueID=1". This fixes
the links to parse the ID from the URL, so that issues can now
be loaded.