[Imaging Browser] Hide Sequence Type in data table (#4797)
The imaging browser data table had a "Sequence Type" column
added by #4692.

This was unintentional, as the data is only there for the
filter. The table does not properly format/display it and
will likely not scale to large protocols.

This sets the column back to hidden, but the filter should
still work.