[Issue Tracker] Fix "Watching" filter and duplicated rows in issue tracker (#4815)
The query for the issue tracker was joining the issues_watching table for
the watching filter, but not restricting the join to the current user. This
resulted in the row being duplicated n times for the number of watchers,
and the "Watching?" filter returning rows if there were any watchers, not
only the current user.

Fixes #4810, #4813.