[Media] Fix HTTP response codes when uploading files
All types of errors were being returned to the user as
internal server error. This was clearly incorrect in
some situations where the error was caused by user error.

Update the showMediaError to take a status code to go
with it, and at the same time improved some error messages
(such as permission denied, which didn't include any
message at all before this change.)