[Travis] Update Dockerfile to use MariaDB instead of MySQL
The base docker image for PHP was upgraded to add support
for Debian Buster 6 days ago. Debian Buster doesn't ship with
MySQL, but instead uses the (compatible) MariaDB client. This
updates the docker file to install mariadb-client instead of
mysql-client, and fix the Travis errors that are currently
happening when building the LORIS docker image.

See also: https://stackoverflow.com/questions/57048428/e-package-mysql-client-has-no-installation-candidate-in-php-fpm-image-build-u,
https://github.com/docker-library/php/commit/ccbe53516fa2fd2d3b0ba680b83148b443a6ccf1