[Data Dict] Update Data Dictionary module to use Data Framework
This updates the datadict module to use the data framework, using
the DataFrameworkMenu class infrastructure that was put in place
in #4918.

This necessitated updating the datadict from an older style of
our React components (where the QuickForm was serialized to
javascript server side) to the a newer form (where most of
the display concerns are handled from React, except for the
dynamic parts which must be sent by implementing getFieldOptions).

Unfortunately, the current React side filtering doesn't support
"Keyword" search in the same manner that was used, so it's
replaced by more consistent/traditional filters for now.