[Design] Add interface for User objects
This adds an interface to describe a User and essential
properties that something which implements User must have.

I started with the User class, removed anything related
to password/permissions (which is a different concern from
describing a User), and removed the functions that were
variations of getSites(), then added Active and Pending
(which are in the User Accounts menu filter.) This is
what remained.

Note that the User class doesn't currently implement this
interface.