[Config] Add settingEnabled() helper function (#4940)
Add a new helper function settingEnabled to determine whether a boolean-style config setting is enabled or not.

This is designed to add some syntactic sugar and hide some of the implementation details of our Config library from calling code.

The calling code only cares whether a setting is enabled. Developers shouldn't need to concern themselves with doing a strict string type comparison on config settings.

e.g. this code:

if ($config->getSetting('useEDC') === 'true') {
    ...
}

can now be written as

if ($config->settingEnabled('useEDC')) {
    ...
}