[Tools] Add script for changing config settings (#4973)
This adds a new script to change the config settings
in the database without SQL access. This may be is useful
when the Config module page is unavailable due to bad
paths or otherwise when providing user support, where
it can be easier to just tell the user what to type
copy/paste in a terminal than describe where to find
a setting on the frontend.