[candidate_parameter] Consent Status data validation bugfix (#5000)
This fixes the logic for data validation on consent status tab. As it stands, if no status stays no or record existed as NULL the validation checks that :

- consent date is not empty
- withdrawal date is empty

However, if Consent Status at one point changes from 'YES' to 'NO' ==> user must fill in date of withdrawal. Therefore, consents that = 'NO' and have non-empty dates of withdrawal are considered non-valid (Displays error message).