Add new phan rule PhanUndeclaredVariableDim to static analysis (#4917)
This adds a rule to make sure that when you access an array that the key you're using makes sense. E.g. if you are accessing an element from an array using a non-numeric key, but the array is actually declared to be a string, this will flag that case.

It fixes all issues identified by phan.