Update readme.md
Removed "The Hard Way" from the instructions based on comments from @driusan - "It's in the README of the aces/Data-Query-Tool repo, and only required for developers. The code that you need to know how to do that for also hasn't been touched in years." Also deleted the Erica setup since that is only required for "the hard way". Fixed formatting based on comments from @johnsaigle