[Data Tables] Improve efficiency of DataTable filtering for large data tables
This improves the efficiency of the DataTable JSX component when dealing with
large data tables. When counting the number of rows which match the filters
for the count in the header/footer, it returns an array of indexes which matched
the filter rather than returning a count. This prevents the need to needlessly
duplicate the work of tracking which rows match for the data table.

The indexed array is then sorted, and we work with that sorted list of indexes
instead of repeatedly going over this.props.data after the first time the matching
rows are calculated.

I suspect that somewhere in the deleted code there was an O(n^2) iteration over
the data that was avoided by re-writing and only directly working with the filtered
data after computing it, but I didn't track down where that was happening to it's
possible there was some other cause.