[Utility] Add isPositiveInteger function (+ unit tests) (#4425)
When checking data coming from e.g. the database or a form, there are some weird subtleties with using intval(). It'll return 0 on an input of null which makes it hard to differentiate between inputs that are null and 0 and those that are unset. It also has non-intuitive behaviour for array and object inputs, see: https://secure.php.net/manual/en/function.intval.php

This PR adds a simple function to check whether a given value is a positive integer as well as unit tests to prove that it works.